dlm_controld: avoid full plock unlock when no resource exists
When a plock unlock operation is read, we can immediately return if
the resource does not exist.  If there were any locks to unlock, the
resource would exist.  This optimization avoids creating/caching the
resource and sending the unlock message unnecessarily.

(When a process holding an *flock* exits without unlocking it, the
vfs cleanup code sends us a full plock unlock on the file, even if
the process has never used plocks.  This means we can potentially
see a lot of plock activity for processes that never touch plocks.
This patch reduces it, although it would be nice to detect that a
process has no plocks earlier, while still in the kernel.)

bz 575103

Signed-off-by: David Teigland <teigland@redhat.com>