dlm_controld: add plock checkpoint signatures
A signature of the plock checkpoint data is made by the writer and
included in the "checkpoint ready" message it sends.  The reader
then computes the signature of the data it reads from the checkpoint
and compares against the signature of the writer.  If they don't
match, the reader has retrieved incorrect plock state so it disables
plock operations for the given lockspace on that node.

bz 578625

Signed-off-by: David Teigland <teigland@redhat.com>