dlm_controld: fix plock checkpoint signatures
Commit e2ccbf90543cf1d163d1a067bf5a8ce049a9c134 for bz 578625
was not correct to use "p_count" (a count of plocks) in the
signature calculation.  When plock_ownership is on, the plocks
under an owned resource are not copied into the checkpoint.
However, the node writing the checkpoint counts all these
owned plocks and factors the count into the signature.  The
node reading the checkpoint does not get the plocks, so its
count of plocks is different, causing the signature calculation
to be different.  It will then disable plock operations.
It would be very common for this to occur in practice, so the
impact is very high.

bz 618806

Signed-off-by: David Teigland <teigland@redhat.com>