dlm_controld: fix save_plocks initialization
When dlm_controld joins a cpg and begins receiving plock messages, it
needs to save all those plock messages for processing after it initializes
plock state from a checkpoint.  Instead of being initialized to 1,
saved_plocks started as 0 and was initialized to 1 shortly after the join.
This left a short span of time where a plock message could arrive and be
processed immediately instead of saved, which would cause the node's plock
state to be out of sync with the other nodes, which could lead to any
number of different problems with plocks.

bz 623810

Signed-off-by: David Teigland <teigland@redhat.com>